Feat: modify useTodo
