--- src/hooks/useHook.tsx
@@ -1,9 +1,0 @@
-import React, { FC } from 'react';
-
-interface Props {}
-
-const ComponentName: FC<Props> = (props) => {
-  return <div></div>;
-};
-
-export default ComponentName;
